#12 Extract OrganizationIdentityService interface from InseeService
