+++ src/main/kotlin/com/fabien/organisationIdentity/Model.kt
@@ -1,5 +1,7 @@
-package com.fabien.organisationIdentity.insee
+package com.fabien.organisationIdentity
 
+import arrow.core.Either
+import com.fabien.OrganizationIdentityError
 import kotlinx.serialization.Serializable
 
 @Serializable
@@ -19,3 +21,13 @@ data class PaginatedOrganizations(
     val page: Int,
     val total: Int,
 )
+
+fun interface OrganizationIdentityService {
+    suspend fun fetchIdentities(
+        nationalId: String?,
+        searchText: String?,
+        zipCode: String?,
+        pageSize: Int,
+        page: Int,
+    ): Either<OrganizationIdentityError, PaginatedOrganizations>
+}
